refactor : 지원, 취소, 체크
